feat: Finish function to join game, add helper functions for receiving messages
Client can now join a room, or create it if the room doesn't exist.

Added functions to wait for a reply. Might want to change main run function to
make use of some of these.

May want to reconsider the way messages are treated -- received messages are
more like a list of messages.