feat: Restructure message sending, begin message parser
Change structure of code so that as much message receiving is done in the main
'run' loop as possible. This avoids the need for extra functions to handle
finding replies. Client object will now be run as follows:

```
pyclient = client.CurvytronClient('serveraddress')
pyclient.start()
pyclient.join_room('roomname')
```
So have to call a function to join the room after starting the main loop. Could
go further and have to call a function to connect to server after starting teh
main loop -- not sure if that would be better.