feat: Add start game handshake, expand message parser
Client can now join a room, set status to ready and parse enough of the
messages that appear during the game to be useful. There are other
messages that are sent -- some that we don't care about (because we
don't need to show the user e.g. the ready status of other players) and
some that we may want to expand the parser to handle, but the messages
that are parsed at the moment are sufficient to collect information
about the status of the game (player positions, score, round start/end).